+++ nix/package.nix
@@ -8,7 +8,7 @@ let
 
   ui = pkgs.stdenv.mkDerivation (finalAttrs: {
     pname = "trove-ui";
-    version = "0.4.3";
+    version = "0.4.4";
 
     src = cleanSrc;
 
@@ -37,7 +37,7 @@ in
 
 pkgs.buildGoModule {
   pname = "trove";
-  version = "0.4.3";
+  version = "0.4.4";
 
   src = cleanSrc;